boom: expose 'BOOT_ID' key in BootEntry container
Make the BOOT_ID value accessible via the dictionary emulation of
the BootEntry class. This allows the value to be obtained either
via:

  id = be.boot_id

Or:

  id = be[BOOT_ID]

The latter form also permits iteration using the special methods
keys(), values(), and items().

Since all keys are included in the string representation of a
BootEntry object, and since the BOOT_ID is calculated using the
string representation, this needs special handling in order to
avoid recurssion in the BootEntry.boot_id property getter.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>