boom: make BootEntry._bp public
The BootParams class is public and is used to configure a new
BootEntry. Make it a public member (no leading underscore) to
indicate that clients may rely on being able to access it.

Users of the member should still test for None since it is
valid for a BootEntry to not have any associated BootParams.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>