boom: recover BootParams from on-disk BootEntry data
Attempt to reverse the profile templating of on-disk BootEntry
objects when they are read into memory: for each templated key
in the BootEntry, construct a regular expression and attempt to
match it against the read entry data. In the case of a match set
the corresponding BootParams member to the recovered key data.

This simplifies modification of existing on-disk entries and
makes the state of written and unwritten BootEntry objects more
consistent.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>