boom: add Selection interface
Add a generic selection interface to be used by the find_*, list_*,
and print_* interfaces. The 'Selection' class provides a single
object representing match criteria for boom BootEntry, BootParams
and OsProfile objects.

Initially only a simple boolean AND of all set criteria is
supported. This can be extended in future to provide more complex
selection criteria.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>