boom: use $machine_id-$boot_id[:7]-$version.conf as entry pattern
Without something in the entry file name to distinguish between
different boot configurations it is impossible to have two BLS
configuration entries for the same machine ID and kernel version.

Since the $machine_id-$version.conf name is only a suggestion add
the truncated boot_id to the generated name to avoid collisions.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>