boom: allow key regex with embedded spaces
The %{root_opts} format key is special in that it is applied
recursively: expanding to the templated value of ROOT_OPTS_LVM2,
ROOT_OPTS_BTRFS, or both.

Allow the key regex generator to generate regular expressions
with embedded spaces, to allow each fragment to be captured
separately, and maintain a table of format keys and the number
of spaces they can contain.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>