boom: rework boom path configuration
Using module members to pass path configuration between the main
boom module and its submodules is fragile: each import evaluates
the imported constants once and stores the result in its own
namespace, meaning that alterations to the default path must be
made carefully to ensure that all imported modules are in sync.

Replace the constants in boom (BOOT_ROOT, BOOM_ROOT) with a pair
of functions and call these from each sub-module to ensure that
the current path values are always used:

  get_boot_path(), set_boot_path()
  get_boom_path(), set_boom_path()

These functions are mainly of use during testing and debugging
and can be used to point the library at an alternate /boot or
/boot/boom location.

To ensure the osprofile and bootloader submodules always use
the correct configured paths replace the BOOT_ENTRIES_PATH and
BOOM_PROFILES_PATH module variables with functions:

  boom_entries_path()
  boom_profiles_path()

That are used throught the modules and which can be imported
by other modules needing to reference these file system
locations.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>