boom: implement sort keys for BoomReport
Accept a string list of sort keys in the BoomReport constructor,
using the same notation as libdm-report: "[-|+]<fieldname>[,*]".

Fields are sorted in the order they appear in the list. Fields
with no prefix, or a '+' prefix are sorted in ascending order.
Fields with a '-' prefix are sorted in descending order.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>