boom: add 'kernel' and 'initramfs' format keys
Add a format key to allow the 'linux' and 'initrd' properties
of a BootEntry to be used in format strings. Each must only
be used in format strings other than the one that defines the
key to avoid recursion in _apply_format().

This is needed since the Fedora 26 kernel options template
embeds the boot image file as BOOT_IMAGE=/path/to at the
start of the command line.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>