boom: fix BootEntry.__from_data pop-if-set logic
The BootEntry __from_data builder should clear members of the
supplied entry_data dictionary only if they match the value that
would be generated by the configured OsProfile and BootParams.

If the value differs it has either been hand-edited, or values
from the template were overridden when it was created: maintain
the edited values for display and modification.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>