boom: allow /dev/vg/lv notation for --root-lv
Allow the use of the /dev prefix when specifying an LVM2 logical
volume (permitted by the LVM2 tools) and automatically remove
it when specifying rd.lvm.lv.

Also attempt to check that the LV name is in a valid format and
generate an error if the --rootlv and --rootdevice arguments do
not match.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>