boom: attempt to parse LV name from --root-device
If --root-device is given, but --root-lv is not, attempt to parse
the LVM2 logical volume name from the device string passed for
--root-device. If this fails it is a non-LVM device and no error
is reported (unless it fails the device presence check).
Otherwise the parsed name is taken as the value of --root-lv and
passed to the rd.lvm.lv parameter in generated boot entries.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>