boom: add legacy bootloader infrastructure
Add infrastructure for handling legacy boot loader configuration
syntax: this adds a new submodule, 'boom.legacy' that contains
classes and functions for working with legacy formats.

Legacy boot loader notation is supported by means of decorator
classes that wrap BootEntry and alter the format of the data
returned by the __str__() method (which returns BLS notation for
BootEntry objects).

When legacy formats are used, BLS remains the canonical store for
all boom boot entries: the legacy support is write-only and the
entire set of entries is synchronised to the legacy configuration
file each time.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>