boom: add persistent configuration infrastructure
Add a 'config' submodule to the 'boom' package to manage an on-disk
persistent configuration.

A new class, BoomConfig, is defined in the 'boom' package to carry
configuration values used throughout boom (currently global path
locations and the configuration of legacy bootloader support).

The 'config' submodule provides functions to read and write this
configuration and to validate the data retrieved.

API users may interact directly with the configuration object to
modify the active configuration or to retrieve values.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>