bootloader: allow appending and dropping of boot options
Add the facility to append extra, non-template options to the
kernel command line, as well as the ability to remove selected
options that are specified by the OsProfile template.

New 'add_opts' and 'del_opts' kwargs are added to BootParams
initialiser and reflected in the 'add_opts' and 'del_opts'
members. These accept a list of strings representing the
options to add and remove respectively.

When a BootEntry is read from disk, appended options are
recovered from the on-disk options line (any option present
not templated by the relevant OsProfile is assumed to have
been appended), and any expected options from the template
that are not found are assumed to have been dropped.

This allows options such as 'debug' or 'noacpi' to be added to
a single BootEntry without modifying the OsProfile or making
post-hoc edits to the generated BLS file.

In addition, option such as "rhgb" and "quiet" that are part
of the default options for some OsProfile instances can now
be selectively dropped when creating a new BootEntry.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>