bootloader: rename all BOOT_* key contants to BOOM_ENTRY_*
The key prefix for the boot entry keys has been "BOOT_" until now
(for e.g. BOOT_ID). This isn't ideal since these are exported by
the module: other key constants all have a "BOOM_*" prefix.

Since names like BOOM_ID would be confusing (especially in the
HostProfile keyset since it is a combination of keys from the
host class as well as OsProfile), rename these problem keys, e.g.:

  BOOT_ID -> BOOM_ENTRY_BOOT_ID

Unfortunately this makes key names and lists more verbose but it
also clarifies their use and avoids name conflicts when using a
wildcard import of the containing module.

This gives us a consistent set of key prefixes over the three
types of container object boom exports:

  BOOM_ENTRY_* -> BootEntry
  BOOM_OS_* -> OsProfile
  BOOM_HOST_* -> HostProfile

Currently key constants are not used anywhere outside of the
boom.bootloader and tests.bootloader_tests modules so the change
is fairly self-contained.

External API users will need to update to the new names but there
are expected to be few if any (and the interfaces have not been
declared as stable as of boom-0.9).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>