osprofile: rename 'name', 'short_name', 'version', and 'version_id'
Rename OsProfile properties that may collide with other fields in
the HostProfile aggregate class.

Since a host has "name" and "short_name" properties (the FQDN and
hostname respectively) this avoids conflict between fields that
are unique to the HostProfile and those that map to its embedded
OsProfile.

This also better reflects the key names used for dicationary-style
access to the OsProfile container and is consistent with the names
of Selection class members.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>