bootloader: fix option deletion for overridden be.options
If a BootEntry's options attribute is overwritten (i.e. a value
is assigned to be.options, causing it to be stored in the entry's
_entry_data dictionary, and to shadow the corresponding OsProfile
template), a different path is taken at the end of the be.options
property method (this happens since no key substitution is done on
the overridden options value).

This branch was missing a call to 'del_opts()' to perform option
deletion according to the setting of be.bp.del_opts.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>