bootloader: track dirty status of BootParams objects
Boot parameters are stored separately from BootEntry objects in
an instance of class BootParams. Since this object ('bp') is a
subordinate of the containing BootEntry there is no direct means
of signaling changes in the bp to the containing be (without an
ownership/callback scheme which is undesirable since it prevents
sharing of bp objects between several be instances).

To ensure that a BootEntry is always aware of changes to its bp
(either substitution of the whole object, or a modification to one
or more of its keys), implement a generation counter scheme for
BootParams.

Each BootParams.generation is an integer counter initialised to
0 at object creation. Every attribute modification increments the
counter by 1. When a BootParams object is newly associated with a
BootEntry, the BootEntry records the current generation number.
The generation number stored in the BootEntry is then compared
against the current bp.generation whenever a boot_id for the
object is requested. If the generations do not match, the entry
dirties itself, re-sets the generation number, and generates a
new boot_id accordingly.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>