bootloader: properly clear templated keys on BootEntry load
When a BootEntry is read from disk, it has full values for all
keys, including those that are generated from the template strings
provided by the OsProfile.

This mechanism allows object API users to selectively overwrite
any BootEntry attribute in order to override the template defined
behaviour (and to have those modifications preserved across write/
read operations).

To allow an in memory BootEntry (or attached BootParams) to be
edited, these keys must be removed from the _entry_data dictionary:
pop all matching keys that are present, and that have a value that
is equal to the template-defined string for the current BootParams.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>