command: fix profile resolution
When resolving an OsProfile or a HostProfile for command line
operation, _find_profiles() must take account of either explicit
command line arguments, or an implicit OsProfile (via version tag
match), or HostProfile (via machine_id match).

Ensure that the precedence for profiles is applied correctl, and
that implicit options are checked appropriately before returning
a match.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>