osprofile: add _profile_exists() helper
There is the possibility of a circular dependency in defining the
null profile and using find_profiles() from within the OsProfile
loading routines: if we try to check for the null profile while
loading, by using find_profiles(), it will find that profiles are
not loaded, and attempt to load them by calling load_profiles().

Avoid the circularity by defining a separate, simple existence
test that scans the list of loaded profiles for a matching OS
identifier.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>