bootloader: add drop_entries()
Add a method to explicitly drop the list of in-memory BootEntry
objects. This allows users of the API to abandon uncommitted
changes and to re-set the list to a known state.

This is currently used by the test suite fixture methods to
ensure a coherent view of sandbox test data.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>