boon: move {get,set}_boom_config_path() out of boom.config
Since the configuration path is global state, move the set/get
functions for the boom configuration path (boom.conf) to the main
boom module. This means that the other path setting functions can
automatically update the configuration path when the path to the
/boot or /boot/boom directories is changed, without running into
problems with recursive imports.

This fixes problems found with path consistency when adding test
coverage for the boom.config module.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>