boom.legacy: remove dupe _legacy_format_error() definition (LGTM)
The _legacy_format_error() wrapper is a helper to raise the
BoomLegacyFormatError() exception, and perform message formatting
and temporary file clean up.

It is only required for the clear_legacy_format() case (which
must search for the Boom begin/end tags, and remove Boom defined
boot entries found in this region of the file).

Remove the unused definition from write_legacy_format().

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>