tests: add predicates and functions for dynamically finding LVs
Add a have_lvm() predicate that can be used to skip tests that
require a real logical volume path, and functions to discover
a suitable usable LV name and path on a system that has at least
one configured LVM volume group.

For testing we only require that the device exist: they are never
modified or accessed from the test suite.

Related: #28

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>