boom.legacy: treat legacy format sync errors as non-fatal
While it is an error to fail to write the legacy bootloader config,
it should not be fatal: the user can always fix the problem that is
preventing the sync, and then run 'boom legacy sync' to update the
configuration files.

This prevents hard errors on systems that have been mis-configured,
for e.g. a system with legacy_enable and legacy_sync set, but where
no grub1 installation exists.

Related: #31

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>