boom.bootloader: support Grub2 environment variables in options
Support the presence of Grub2 environment variables in BootEntry
option strings:

  $kernelopts $tunedopts

These are expanded by the bootloader on boot to values stored in
the Grub2 environment block in /boot.

Values obtained from the bootloader environment are treated as
though they were defined by the corresponding OsProfile template
and are not treated as having been appended to the BootEntry
when recovering boot parameters in BootParams.from_entry().

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>