boom.osprofile: relax matching of FMT_KERNEL and FMT_INITRAMFS
The kernel and initramfs image format keys are never used to parse
boot parameters from on-disk entries: since these values may be
overridden to non-profile derived strings on the command line (e.g.
for testing or debugging), and since the scope of the value is well
defined it does not make sense to require an exact match to the OS
profile template string, for e.g. '/vmlinuz-(\S+)': just match a
general regular expression that will accept any possible value of
the format key, including values resulting from overriding the
template defined values on the command line.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>