boom: add BootEntry.entry_path
Add a public property to determine the on-disk location of a
BootEntry object and use it to report the entry path from reports
in boom.command.

For boom managed entries this is the path to which boom would
write the entry file (including the boot_id in the file name). For
foreign entries it is the path from which the BootEntry was read.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>