boom.bootloader: rename BOOM_ENTRY_ID to BOOM_ENTRY_GRUB_ID
The 'id' BLS key is defined by the Red Hat BLS implementation used
in Red Hat, Fedora and CentOS distributions. It corresponds to the
Grub2 menuentry --id option and influences the sort ranking of
entries displayed by the system bootloader.

Rename the Boom key constant to make this relationship clear: the
BootEntry property must be named 'id' (since its value is set using
the Boom to BLS name mapping), but the name of the constant can
reflect the fact that this is a non-standard BLS key.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>