boom.command: fix --uname-pattern heuristic for Fedora
Default uname patterns for Fedora are a little too broad:

  # boom create --from-host
  ...
    Optional keys: "grub_users grub_arg grub_class id", UTS release pattern: "fc3"
                                                                             ^^^^^
The heuristic used to automatically generate --uname-pattern for
known distributions incorrectly uses version_id[0] to set the
numeric part of the pattern for all Red Hat-like distros. This is
to accomodate the RHEL "elX_Y" minor release naming convention.

Instead explicitly check for the presence of "_" and return the
string with "_" and the minor version suffix stripped.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>