boom.cache: add _is_restored() test
Add a function to test whether a boot_path is the original file,
or if it was restored by boom. Initially rely on testing for the
presence of a dot-file named for the restored image file, e.g.:

  /vmlinuz -> /.vmlinuz.boomrestored

As this will work with any VFS type. This can be extended to use
xattrs where the /boot file system supports them.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>