boom.cache: add CacheEntry implementation
Add a CacheEntry class to represent entries in the boom boot image
cache. This class is a lightweight wrapper around the cache state
that is stored in the private dictionaries defined in boom.cache.

A CacheEntry maps a cached path (e.g. "/vmlinuz") to a set of path
metadata (mode, ownership etc.) and one or more cached images.
Images are identified by a SHA1 image identifier (the SHA1 hash
of the file content) and each image identifier is stored along
with a corresponding timestamp indicating the mtime of the file at
the time the image was added to the cache.

Users of the cache will operate on cache entries by calling methods
of instances of the CacheEntry class.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>