boom.command: fix command validation regression (0bde1513)
This is a regression due to commit 0bde1513 ("boom: don't allow cache
commands if !config.cache_enable"). This adds a check to prevent a
command of type "boom cache" running when the cache is disabled in
boom.conf. The check is applied before validation of the parsed
cmd_type, leading to an unhandled exception in the event that the
provided command argument is invalid.

The fix is to move the validation before the cache check and this commit
also adds additional checks to distinguish between an unknown subcommand
with default type (entry) and a missing argument, which both previously
resulted in the same terse "boom: error: too few arguments" message.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>