boom: allow expand for all BootEntry keys in BootEntry.__str()
Initially bootloader variable expansion was limited to the options
key of BootEntry options. The use of Grub2 bootloader variables
has since expanded to the optional grub_* keys as well as the
initramfs image key (and could potentially be used in others in
future). Remove the special treatment of the options key and allow
expansion of all BootEntry key values in string representations.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>