tests: add infrastructure for testing --expand-variables
The code that backs the --expand-variables option relies on a call
to the grub2-editenv command to obtain the Grub2 environment block.

Add a mock version of this command in a new tests/bin directory and
prepend this path to the PATH environment variable in a new tests
module helper function set_mock_path()

Test classes that wish to use this facility should call this
function from their setUp() method.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>