tests: consolidate RHEL8 OS profiles
Since commit 0e19265 there are two sample RHEL8 OsProfiles in the boom
release:

  examples/profiles/f44fb528ff8360ad67e2fe0274750b838da0bd6a-rhel8.1.profile
  examples/profiles/1ba00433fd7ba24532ec42c3cdf9f457821e6e02-rhel8.2.profile

OsID    Name                     OsVersion
f44fb52 Red Hat Enterprise Linux 8.1 (Ootpa)
1ba0043 Red Hat Enterprise Linux 8.2 (Ootpa)

These are generated from the respective /etc/os-release files for the
corresponding release.

There's no real need to do this unless a user is actually dealing with
multiple minor versions in parallel AND needs to specify different
options or other template values for the two releases.

We support the ability to have differentiated profiles for advanced use
cases where this may be needed but for most users this is an unnecessary
distraction and may lead to confusion (for example when making edits to
a profile) since there is not yet a reliable mechanism for determining a
minor release by uname.

Consolidate the 8.1 and 8.2 profiles into a single rhel8 OsProfile:
users who actually work with both releases can create the extra profile
manually and will need to specify --profile in order to unambiguously
select the active profile.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>