boom.cache: ignore foreign boot entries when reference counting
Do not include non-boom managed boot entries when determining the use
count of cache entries: this leads to a false warning when creating
entries with --backup since the origin path is added to the cache,
restored to the backup path, and then removed from the cache.

A better fix would be to move the implementation of --backup from
boom.command to a new boom.cache.backup_path() API function but
this is a self-contained change in the CacheEntry.count method
that avoids interference from system managed boot entries.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>