boom: move backup logic into boom.cache
The logic for handling backup images was originally implemented
in the boom.command module: this allows it to be used by the
boom CLI without committing to offering it as a supported API.

Backing up an image in this model is handled by a sequence of
cache operations:

  - Generate a backup name
  - cache_path() for the original path
  - CacheEntry.restore() for the backup path
  - uncache_path() for the original path

This has the effect of adding the backup path to the cache without
the image having to exist at the time (i.e. without the command
module needing to perform any file operations).

Unfortunately this causes side-effects when other entries also
use the same original path:

 * Calling uncache_path() originally unconditionally removed
   the path from the cache.

 * This triggers a warning that an in-use cache entry is being
   removed if another entry uses the same image path.

Fixing this in boom.command would require "peeking" into the
cache state to know whether the path was cached or not before
the operation (and therefore whether it should be uncached or
just left at the end).

This is pulling a bit too much of the cache internals over to the
command module, so move the logic out of the internal helper at
boom.command._cache_path() and into a new module level function
boom.cache.backup_path() that accepts the original and backup
paths as arguments.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>