tests: update test host profiles
Two of the test host profiles included in tests/boom/hosts were
created using an early development version of the HostProfile
class. The original version generated different HostID values for
these profiles, causing the file names to change on-disk if the
profiles are re-written (for e.g. by a call to load_host_profiles()
followed by a call to write_host_profiles()).

The change does not affect any host profiles created with released
versions of boom since the previous HostID computation was never
included in a release.

Re-sync the test data so that the values on-disk reflect the in
memory state of the objects.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>