boom.bootloader: mark entries with mismatched filename read-only
A BootEntry loaded from a file that does not conform to the boom
file name structure convention (machine_id-boot_id-version.conf)
should be treated as a foreign entry and marked read-only.

This prevents boom from modifying entries that have been hand-
edited or that were created by a different tool (e.g. the rescue
entries present on Fedora and RHEL systems that do not contain
a valid version string in the BLS 'version' key).

Entries for which no matching OsProfile can be found are already
hidden from the user (since we cannot manipulate them correctly):
this change extends read-only marks to entries with valid version
and other parameters but where the boot_id contained in the file
name does not match the content stored in the file.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>