boom.osprofile: avoid duplicate null OsProfile
The null OsProfile exists to pair with any BootEntry for which a
valid profile cannot be found (it specifies no templates and uses
the raw values parsed from the BLS snippet for any entries that
are matched to it).

The drop_profiles() function is responsible for instantiating the
null profile: calling the OsProfile initialiser will add the new
profile to the global _profiles list. Since the function also calls
_profiles.append() for the new profile the entry is duplicated in
the list. Drop the call to append() and ensure that only a single
instance is present in the list.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>