tests: skip test_write_host_profiles_fail if have_root()
The test_write_host_profiles_fail test relies on the fact that
DAC permissions will prevent host profiles from being successfully
written. Since DAC checks are not applied to root this subverts
the point of the test and leads to pollution of /boot/boom/hosts
with test artefacts (the set of host profiles loaded at the time
of the test).

Skip this test if run as root.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>