boom: fix precedence and handle conflicts when merging options
The --add-opts and --del-opts options specify lists of options to
add or remove from the default set when creating, editing or
cloning a boot entry.

When editing or cloning the command line values should be merged
with the existing values in the BootEntry's attached BootParams:
command line values should override any changes present in the
original entry (so a --del-opt cancels a prior --add-opts and
vice versa).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>