boom.bootloader: remove add_opts from stored BootEntry.options
Options that are not specified by a BootEntry's OsProfile template
are tracked in the corresponding BootParam.add_opts list and are
automatically added to the stored (template) option string when
the BootEntry is formatted as a string.

To prevent add_opts being displayed twice when printing a BootEntry
they must be removed from the stored options string during loading.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>