boom.bootloader: do not set BootParams attrs for anonymous option words
Fixed strings that appear in the options template and that do not
have a corresponding value (for e.g. "ro") are "anonymous": there
is no corresponding BootParams attr name to be set.

Currently this causes BootParams.from_entry() to attempt to:

  setattr(bp, "", "ro")

Although this silently succeeds it is clearly wrong: only attempt
to set an attribute value when we are dealing with a named attr.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>