boom.bootloader: do not set matches[] until after lvm_root_lv check
When we see an rd.lvm.lv option string we cannot set the matches
member for the option to True until we have verified that this
rd.lvm.lv instance corresponds to the root device: move the assign
to matches[word] after the _match_root_lv() check so that other
LVs (for e.g. swap brought up in the initramfs for resume) are
properly ignored.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>