boom.command: do not use set() to dedupe option lists
Avoid using set() to deduplicate the list of options when cloning
or editing entries since the order of the results may change from
one run to another (with the same input data) due to the hashing
carried out by the set data type.

This causes the boot_id to change unnecessarily due to option
shuffling when performing two subsequent edits, where the second
edit reverts the first.

Manually filter the list entries when merging existing and command
line option values and ensure that the order of each is preserved.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>