boom.bootloader: allow empty BLS keys
The BLS format allows empty keys (i.e. a line containing a valid
key name, but no subsequent whitespace separated value string).

A practical case where this occurs is when cloning a system defined
entry with --expand: in this case, references to Grub2 environment
variables that are empty (for e.g. $grub_users) are replaced with the
empty string.

Modify parse_name_value() to optionally allow empty values and use
this in BootEntry.__from_file().

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>