boom.profile: allow optionally selecting the null profile
Allow find_* functions to optionally select the null profile (or
BootEntry objects that have the null profile as their OsProfile
etc.).

This allows otherwise valid BootEntry objects that use the null
profile to be displayed and cloned (for e.g. the Fedora rescure
boot entry, that uses the string "0-rescue-%{machine_id}" as the
version).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>