boom.bootloader: handle machine_id for Red Hat native BLS entries
Red Hat's native BLS snippets do not set the machine-id BLS key.
This is not a problem when loading and displaying these entries,
but the BOOM_ENTRY_MACHINE_ID value forms part of the filename
for boom entries that are cloned from these entries.

Handle this by setting the machine_id key for these entries from
the filename, and set a flag on the entry to suppress the output
of the key and its use in calculating the boot_id.

This maintains line-for-line equality between the BLS snippet
and boom output for these entries, but gives normal boom handling
of the machine_id key for any entries cloned from them since the
_suppress_machine_id flag is not inherited by the clone.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>