legacy: sort legacy boot entries when writing configuration
The order of entries in a legacy boot loader configuration file
written by boom.legacy.write_legacy_loader() depends only on the
order returned by find_entries(). This list in turn is populated
in the order returned by the file system via os.listdir().

This means that any change in the order of results returned by
the fs (for e.g. due to an entry file being unlinked and re-written
with a new inode number) is reflected in the generated grub.conf.

This leads to unnecessary churn in the tests/ data: avoid the
problem (and give a slightly more predictable experience for users
of legacy loaders) by sorting the entries list by version before
writing it out.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>