command: fix optional BLS key to command line argument decoding
When reporting errors relating to the use of optional BLS keys
(i.e. --grub-users, --grub-class, --grub-arg, --grub-id) the boom
command module translates the symbolic name of the key to a user
level command line argument, for e.g.:

  Profile with os_id='4abe4f7' does not support --grub-id

This is broken in boom-1.2 since the table is missing the entry
for the --grub-id argument, and the _set_optional_key_defaults()
function that calls it passes the incorrect form of the key
identifier (the BLS key name, rather than the boom API name for
the key).

Fix the table and the call, and add a test case to confirm that
the proper value is returned when decoding optional key names.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>