boom: remove BOOM_ENTRY_OPTIONS hack from BootEntry.__from_data()
Commit 92c2bce3 moved logic for accounting for add/del opts from
the BootParams.from_entry() method into BootEntry.__from_data().
Although this is the more correct location to make these changes
the rewriting of the stored BOOM_ENTRY_OPTIONS value is actually
redundant: the entire value will be cleared by the pop-if-set
logic applied to all template-generated keys in the final stage
of loading a BootEntry from an entry_data array.

Additionally the logic conflicts with the handling of host profile
options modifications causing entries with an attached HostProfile
to unexpectedly change boot_id.

Remove the special handling of BOOM_ENTRY_OPTIONS and simplify the
BootEntry.__from_data() method.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>