Revert "docs: rename toctree document to "main_doc""
This reverts commit 86120d06efa47538c62db1a3f081343806404f70.

The original commit attempted to implement the suggestions of the
Conscious Language Initiative:

  https://github.com/conscious-lang/conscious-lang-docs/blob/main/recommendations.md

Specifically by renaming the toctree document element from the name
'master_doc' to 'main_doc'. Unfortunately this change is incompatible
with older versions of the Sphinx toolchain leading to an error when
building the package documentation:

  Sphinx error:
  master file /builddir/build/BUILD/boom-1.4/doc/contents.rst not found
  make: Leaving directory '/builddir/build/BUILD/boom-1.4/doc'
  make: *** [Makefile:53: html] Error 2

(when built with python3-sphinx-1:1.7.6-2.el8).

Sphinx 4.0.0 has renamed the directive to "root_doc" and this can be
used once this release is available in distributions.