boom.bootloader: initialise _last_path before parsing BootEntry
The path from which a boot entry was read can be useful context for
logging messages, particularly when a boot entry has been corrupted or
manually edited (so that the boot_id no longer matches the expected
value).

Move the initialisation of this member before parsing the boot entry, so
that the value is available for log messages.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>