Revert "boom.cache: ignore foreign boot entries when reference counting"
This reverts commit f9704eea7b973863ea5a2bf6ad13cd37abce64f6.

The reason for ignoring foreign boot entries when determining reference
counts no longer exists (a spurious warning when creating entries with
--backup).

This commit causes problems if a boom-managed boot entry is modified
outside of boom's control (e.g. by kdumctl/grubby appending a modified
crashkernel argument): since boom sees the entry as foreign the images
used by it end up with a reference count of zero and are automatically
removed.

Revert the above commit to prevent this behaviour.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>