dist: drop snapshot-remount-fs generator
Since systemd-v255 the systemd-remount-fs service will refuse to run if
the fstab=no kernel command line parameter is present:

    commit a38de31c8b0dc5a5b1b534e766cb12988a2db5d9
    Author: Yu Watanabe <watanabe.yu+github@gmail.com>
    Date:   Wed Aug 9 02:02:23 2023 +0900

        remount-fs: refuse to remount based on fstab when fstab=no kernel command line option specified

        Otherwise, if for some reasons remount-fs is invoked even when fstab=no is
        specified, mounts may get unexpected options from fstab.

        For safety, let's parse the kernel command line option.

This makes the snapshot-remount-fs script (that enables
systemd-remount-fs when booting a snapshot) useless. Remove it from the
package.

In the specific niche case that we were using the service for the
results are correct: we /want/ the fstab options to be applied to the
snapshot root volume but upstream has decided that it is unsafe for the
general case which disables our use of it from snapshot-remount-fs.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>