boom: fix --backup of modified images
If an image has already been backed up by boom and is subsequently
modified attempting to create a new boot entry using the same image
and --backup fails with:

  # boom create --title BackupTest2 --backup --root-lv fedora/root
  ERROR - Could not cache path /initramfs-6.8.9-300.fc40.x86_64.img: Restore failed: CacheEntry state is not MISSING or RESTORED
  Restore failed: CacheEntry state is not MISSING or RESTORED

This can happen when package scripts regenerate the initramfs image
between the time the first and second boot entries are created.

The status of the cached image is reported as CACHED rather than
RESTORED because the hash of the image in /boot does not match the
expected cached value:

  # boom cache list
  Path                                       ImageID Timestamp  State
  /initramfs-6.8.9-300.fc40.x86_64.img.boom0 bcc364b 1723557576 CACHED
  /vmlinuz-6.8.9-300.fc40.x86_64.boom0       10c480b 1714608000 RESTORED

This happens because the logic to generate a path name for the backup
image fails to ensure uniqueness in the case that the image hash has
changed. The original version remains in /boot and prevents the updated
image from being restored.

Fix this my moving the generation of the unique path from boom.command
into boom.cache where we can whether the image is already cached, or if
it is a modified/updated version of the same path name.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>