boom.cache: default to re-using cached images
Add a new 'update' parameter to boom.cache.backup_path() and default to
re-using an existing backup image if one is present.

This protects against situations where the initramfs may be modified in
between creating a snapshot and rollback boot entry: this happens during
updates from RHEL8.10 to RHEL9 where the RHEL8 initramfs is re-built
during the upgrade transaction using the RHEL9 dracut and initramfs
userspace.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>